norminette fix, found a fix to pipex but not really happy with it
