+++ srcs/built-ins/mini_env.c
@@ -1,29 +1,23 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   mini_aux.c                                         :+:      :+:    :+:   */
+/*   mini_env.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: dpaes-so <dpaes-so@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/03/25 16:45:21 by dpaes-so          #+#    #+#             */
-/*   Updated: 2025/04/14 16:19:31 by dpaes-so         ###   ########.fr       */
+/*   Created: 2025/04/14 15:31:00 by dpaes-so          #+#    #+#             */
+/*   Updated: 2025/04/14 16:07:05 by dpaes-so         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../incs/mini_header.h"
+#include "../../incs/mini_header.h"
 
-void	freetrix(char **matrix)
+int	build_env(t_mini *mini)
 {
 	int	i;
 
-	i = 0;
-	if (!matrix)
-		return ;
-	while (matrix[i])
-	{
-		free(matrix[i]);
-		i++;
-	}
-	if (!matrix)
-		free(matrix);
+	i = -1;
+	while (mini->env->my_env[++i])
+		ft_printf("%s\n", mini->env->my_env[i]);
+	return (1);
 }
